refactor: move config to shared folder
