+++ src/app/intro/shared/intro.config.ts
